requirements.txt
aiogram==3.20.0
SQLAlchemy[asyncio]==2.0.28
alembic==1.13.1
pydantic==2.11.4
pydantic-settings==2.2.1
APScheduler==3.10.4
asyncpg==0.29.0
python-dotenv==1.0.0
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
factory-boy==3.3.0
aiosqlite==0.20.0